+++ gamevault/UserControls/SettingsUserControl.xaml.cs
@@ -17,6 +17,8 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Collections.Generic;
 using System.Windows.Markup;
+using gamevault.Helper.Integrations;
+using AngleSharp.Dom;
 
 namespace gamevault.UserControls
 {
@@ -458,7 +460,7 @@ namespace gamevault.UserControls
             {
                 if (File.Exists(((ThemeItem)uiCbTheme.SelectedItem)?.Path))
                 {
-                    File.Delete(((ThemeItem)uiCbTheme.SelectedItem).Path);                   
+                    File.Delete(((ThemeItem)uiCbTheme.SelectedItem).Path);
                     uiCbTheme.SelectedIndex = 0;
                     LoadThemes();
                 }
@@ -539,7 +541,28 @@ namespace gamevault.UserControls
                 ViewModel.DevModeEnabled = true;
             }
         }
+        private void RemoveCustomCloudSaveManifest_Click(object sender, RoutedEventArgs e)
+        {
+            int index = ViewModel.CustomCloudSaveManifests.IndexOf(((LudusaviManifestEntry)((FrameworkElement)sender).DataContext));
+            if (index >= 0)
+            {
+                ViewModel.CustomCloudSaveManifests.RemoveAt(index);
+            }
+        }
+        private void AddCustomCloudSaveManifest_Click(object sender, RoutedEventArgs e)
+        {
+            ViewModel.CustomCloudSaveManifests.Add(new LudusaviManifestEntry());
+        }
 
-
+        private void SaveCustomCloudSaveManifests_Click(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                string result = string.Join(";", ViewModel.CustomCloudSaveManifests.Where(entry => !string.IsNullOrWhiteSpace(entry.Uri)).Select(entry => entry.Uri));
+                Preferences.Set(AppConfigKey.CustomLudusaviManifests, result, AppFilePath.UserFile);
+            }
+            catch { }
+            MainWindowViewModel.Instance.AppBarText = "Successfully saved custom Ludusavi Manifests";
+        }
     }
 }